marinee_weather_forcasting/requirements.txt
pandas
numpy
requests
scikit-learn
streamlit
joblib
altair
pydeck
python-dotenv
